+++ migrations/topicimage_install.php
@@ -15,7 +15,7 @@ class topicimage_install extends \phpbb\db\migration\migration
 	public static function depends_on()
 	{
 		return [
-			'\phpbb\db\migration\data\v320\v320'
+			'\phpbb\db\migration\data\v330\v330'
 		];
 	}