+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Topic Image Changelog
 
+## 1.0.1 - 2021-05-26
+
+- Added limit for images in topics.
+- Added Cron for deleting and grabbing images.
+- Code updates.
+
 ## 1.0.0 - 2021-02-16
 
 - First release.